requirements.txt
yfinance==0.2.36
pandas==2.2.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
python-dotenv==1.0.0
schedule==1.2.1
numpy==1.26.3